move Screen internal to common
