+++ hackle-android-sdk/src/main/java/io/hackle/android/internal/engagement/Engagement.kt
@@ -1,6 +1,6 @@
 package io.hackle.android.internal.engagement
 
-import io.hackle.android.internal.screen.Screen
+import io.hackle.sdk.common.Screen
 
 internal data class Engagement(
     val screen: Screen,
