+++ hackle-android-sdk/src/test/java/io/hackle/android/internal/bridge/HackleBridgeTest.kt
@@ -8,6 +8,7 @@ import io.hackle.sdk.common.Event
 import io.hackle.sdk.common.HackleRemoteConfig
 import io.hackle.sdk.common.ParameterConfig
 import io.hackle.sdk.common.PropertyOperation
+import io.hackle.sdk.common.Screen
 import io.hackle.sdk.common.User
 import io.hackle.sdk.common.Variation
 import io.hackle.sdk.common.decision.Decision
@@ -1290,7 +1291,7 @@ class HackleBridgeTest {
     
     @Test
     fun `invoke set current screen with parameters`() {
-        every { app.setCurrentScreen(any(), any()) } answers { }
+        every { app.setCurrentScreen(any()) } answers { }
         val parameters = mapOf(
             "screenName" to "mainActivity",
             "className" to "mainActivityClass",
@@ -1299,8 +1300,7 @@ class HackleBridgeTest {
         val result = bridge.invoke(jsonString)
         verify(exactly = 1) {
             app.setCurrentScreen(
-                withArg { assertThat(it, `is`("mainActivity")) },
-                withArg { assertThat(it, `is`("mainActivityClass")) }
+                withArg { assertThat(it, `is`(Screen("mainActivity", "mainActivityClass"))) },
             )
         }
         result.parseJson<BridgeResponse>().apply {
