+++ hackle-android-sdk/src/test/java/io/hackle/android/internal/screen/ScreenTest.kt
@@ -1,6 +1,7 @@
 package io.hackle.android.internal.screen
 
 import android.app.Activity
+import io.hackle.sdk.common.Screen
 import org.junit.Test
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
